move ChipThreadWork to server/java folder as bzbarsky-apple's suggestion because is it not suit for all platforms
